refactor: dynamic post fetch
