+++ src/Pages/Explore/Explore.jsx
@@ -1,8 +1,14 @@
-import Header from "../../Components/Header/Header";
-import Posts from "../../Components/Posts/Posts";
-import LeftPanel from "../../Components/LeftPanel/LeftPanel";
-import Footer from "../../Components/Footer/Footer";
+import { Header, Posts, LeftPanel, Footer } from "../../Components";
+import { usePost } from "../../Context/PostContext/post-context";
+import { useAuth } from "../../Context/AuthContext/auth-context";
 const Explore = () => {
+  const {
+    postState: { posts },
+  } = usePost();
+  const {
+    authState: { user },
+  } = useAuth();
+  const exceptMyPost = posts.filter((item) => item.username !== user.username);
   return (
     <div>
       <Header />
@@ -12,8 +18,18 @@ const Explore = () => {
           <div className="max-w-2xl mx-auto rounded-lg px-3 md:pb-0 pb-14">
             <h1 className="text-2xl font-semibold p-2">Explore</h1>
             <div>
-              <Posts />
-              <Posts />
+              {exceptMyPost.length > 0 ? (
+                exceptMyPost.map((item) => (
+                  <Posts
+                    key={item.id}
+                    content={item.content}
+                    username={item.username}
+                    profile={item.profile}
+                  />
+                ))
+              ) : (
+                <h2>Nothing to Explore</h2>
+              )}
             </div>
           </div>
           <Footer />
